Ceph : Correct the reclaim_reset logic
In case of Ganesha failure, once Ganesha comes back, the associated ceph
client has to reclaim its old status/CAPS. For identifying the ceph
client, an uuid string is formed using first exportid. This will lead to
reclaim failure when one removes the first export.
Instead of relying on exportid, defaulting it to make use of 0001 in the
uuid string.
Also propagting back the ceph side failure (as applicable) to Ganesha,
instead of hiding it.

Change-Id: I978138f1d12399e7e23c5f46e424ccc5363f44bd
Signed-off-by: Sachin Punadikar <sachin.punadikar@ibm.com>